cleanup parsers (#560)
