+++ r2r/base/__init__.py
@@ -1,6 +1,3 @@
-from ..pipeline.ingestion_pipeline import IngestionPipeline
-from ..pipeline.rag_pipeline import RAGPipeline
-from ..pipeline.search_pipeline import SearchPipeline
 from .abstractions.base import AsyncSyncMeta, UserStats, syncable
 from .abstractions.document import (
     DataType,
@@ -52,22 +49,7 @@ from .logging.log_processor import (
     LogProcessor,
 )
 from .logging.run_manager import RunManager, manage_run
-from .parsers import (
-    AsyncParser,
-    AudioParser,
-    CSVParser,
-    DOCXParser,
-    HTMLParser,
-    ImageParser,
-    JSONParser,
-    MarkdownParser,
-    MovieParser,
-    PDFParser,
-    PPTParser,
-    TextParser,
-    XLSXParser,
-)
-from .pipeline.base_pipeline import EvalPipeline, Pipeline
+from .pipeline.base_pipeline import Pipeline
 from .pipes.base_pipe import AsyncPipe, AsyncState, PipeType
 from .providers.embedding_provider import EmbeddingConfig, EmbeddingProvider
 from .providers.eval_provider import EvalConfig, EvalProvider
@@ -136,26 +118,8 @@ __all__ = [
     "Triple",
     "KGExtraction",
     "UserStats",
-    # Parsers
-    "AudioParser",
-    "AsyncParser",
-    "CSVParser",
-    "DOCXParser",
-    "HTMLParser",
-    "ImageParser",
-    "JSONParser",
-    "MarkdownParser",
-    "MovieParser",
-    "PDFParser",
-    "PPTParser",
-    "TextParser",
-    "XLSXParser",
     # Pipelines
     "Pipeline",
-    "EvalPipeline",
-    "IngestionPipeline",
-    "RAGPipeline",
-    "SearchPipeline",
     # Providers
     "EmbeddingConfig",
     "EmbeddingProvider",
