+++ r2r/base/parsers/__init__.py
@@ -1,32 +1,5 @@
 from .base_parser import AsyncParser
-from .parser_impls import (
-    AudioParser,
-    CSVParser,
-    DOCXParser,
-    HTMLParser,
-    ImageParser,
-    JSONParser,
-    MarkdownParser,
-    MovieParser,
-    PDFParser,
-    PPTParser,
-    TextParser,
-    XLSXParser,
-)
 
 __all__ = [
     "AsyncParser",
-    "AudioParser",
-    "CSVParser",
-    "DOCXParser",
-    "HTMLParser",
-    "ImageParser",
-    "JSONParser",
-    "MarkdownParser",
-    "MovieParser",
-    "PDFParser",
-    "PPTParser",
-    "ReductoParser",
-    "TextParser",
-    "XLSXParser",
 ]
