+++ r2r/parsers/__init__.py
@@ -1,0 +1,27 @@
+from .media.audio_parser import AudioParser
+from .media.docx_parser import DOCXParser
+from .media.img_parser import ImageParser
+from .media.movie_parser import MovieParser
+from .media.pdf_parser import PDFParser
+from .media.ppt_parser import PPTParser
+from .structured.csv_parser import CSVParser
+from .structured.json_parser import JSONParser
+from .structured.xlsx_parser import XLSXParser
+from .text.html_parser import HTMLParser
+from .text.md_parser import MDParser
+from .text.text_parser import TextParser
+
+__all__ = [
+    "AudioParser",
+    "DOCXParser",
+    "ImageParser",
+    "MovieParser",
+    "PDFParser",
+    "PPTParser",
+    "MDParser",
+    "HTMLParser",
+    "TextParser",
+    "CSVParser",
+    "JSONParser",
+    "XLSXParser",
+]
