+++ r2r/pipelines/__init__.py
@@ -1,0 +1,5 @@
+from .ingestion_pipeline import IngestionPipeline
+from .rag_pipeline import RAGPipeline
+from .search_pipeline import SearchPipeline
+
+__all__ = ["IngestionPipeline", "SearchPipeline", "RAGPipeline"]
