+++ r2r/pipes/__init__.py
@@ -1,16 +1,16 @@
 from .abstractions.search_pipe import SearchPipe
-from .embedding_pipe import EmbeddingPipe
-from .eval_pipe import EvalPipe
-from .kg_agent_search_pipe import KGAgentSearchPipe
-from .kg_extraction_pipe import KGExtractionPipe
-from .kg_storage_pipe import KGStoragePipe
-from .parsing_pipe import ParsingPipe
-from .query_transform_pipe import QueryTransformPipe
-from .search_rag_pipe import SearchRAGPipe
-from .streaming_rag_pipe import StreamingSearchRAGPipe
-from .vector_search_pipe import VectorSearchPipe
-from .vector_storage_pipe import VectorStoragePipe
-from .web_search_pipe import WebSearchPipe
+from .ingestion.embedding_pipe import EmbeddingPipe
+from .ingestion.kg_extraction_pipe import KGExtractionPipe
+from .ingestion.kg_storage_pipe import KGStoragePipe
+from .ingestion.parsing_pipe import ParsingPipe
+from .ingestion.vector_storage_pipe import VectorStoragePipe
+from .other.eval_pipe import EvalPipe
+from .other.web_search_pipe import WebSearchPipe
+from .retrieval.kg_agent_search_pipe import KGAgentSearchPipe
+from .retrieval.query_transform_pipe import QueryTransformPipe
+from .retrieval.search_rag_pipe import SearchRAGPipe
+from .retrieval.streaming_rag_pipe import StreamingSearchRAGPipe
+from .retrieval.vector_search_pipe import VectorSearchPipe
 
 __all__ = [
     "SearchPipe",
