+++ r2r/pipes/retrieval/kg_agent_search_pipe.py
@@ -12,7 +12,7 @@ from r2r.base import (
     PromptProvider,
 )
 
-from .abstractions.generator_pipe import GeneratorPipe
+from ..abstractions.generator_pipe import GeneratorPipe
 
 logger = logging.getLogger(__name__)
