+++ r2r/pipes/retrieval/streaming_rag_pipe.py
@@ -12,7 +12,7 @@ from r2r.base import (
 )
 from r2r.base.abstractions.llm import GenerationConfig
 
-from .abstractions.generator_pipe import GeneratorPipe
+from ..abstractions.generator_pipe import GeneratorPipe
 from .search_rag_pipe import SearchRAGPipe
 
 logger = logging.getLogger(__name__)
